fix(trace): resolve main rebase conflicts
